Refactor MessageService to simplify constructors
The existing codebase with multiple MessageService constructors has been updated to just have one. Initialization of a private `_enabled` field and a `Sender` property has been added to handle possible null cases of the `sender` variable. The changes also replace the usage of `_sender` with `Sender` for message sending logic.