Refactor FoldersController to use FolderContentDtoHelper
Removed the FoldersControllerHelper class and rearranged its functions amongst the existing classes. The majority of the functionality has been moved to the FolderContentDtoHelper class. The FoldersController now uses the FolderContentDtoHelper to get the content of folders. The GlobalFolderHelper is also used directly from the FoldersController for access to root folder IDs. The changes are aimed to streamline and improve the design and readability of the code.