Add culture change feature in user manager
Added an additional feature in UserManager to allow updating the user's culture. The implementation includes handling potential failures by reverting the changes when an exception is thrown. This function is integrated into various other components, including settings controller and user controller. Updated the authentication process to include the option to change the user culture.