Add order management in file and folder DTOs
The code has been updated to consider the order of files and folders in their respective DTOs. A new class `BreadCrumbsManager` has been added to manage the order property. The 'GetAsync' method in both 'FileDto' and 'FolderDto' classes has been amended to implement the aforementioned changes. The DAOs for 'File' and 'Folder' were altered to record the order in the database.