Refactor encrypt and decrypt methods to async
The primary change in this commit is the conversion of the encryption and decryption methods within the InstanceCrypto class to async methods. Adjustments were also made to other classes and methods that make use of these encryption and decryption methods in order to handle the async nature of the methods. This includes changes in classes such as TextileStyler, LdapHelper, EncryptionLoginProvider, and others. This refactor enhances the code performance by utilizing the advantages of asynchronous programming.