Refactor SecurityDao and perform extensive type changes
The commit includes extensive changes to several components. Primarily, it deals with refactorization of SecurityDao which includes changes to the type complements. Furthermore, multiple modifications are performed in InvitationLinkService, FilesMappingProfile and FilesMappingAction to accommodate generic types (string, int) instead of fixed types. The code has been augmented to handle strings and ints in several sections for seamless data transactions. The changes aim to improve data type flexibility in the code.