Implement MessageSenderService and update messaging system
A new MessageSenderService was implemented and integrated, allowing for more effective handling of event messages. Additionally, updates were made to the existing messaging system - the AddAsync() method was refactored and the ForceSave() method was trimmed and relocated for improved readability and functionality. Some unnecessary code was removed, resulting in overall cleaner and more maintainable codebase.