Refactor GetAvailableTitleAsync method across third party providers
This commit unifies the implementation of the GetAvailableTitleAsync method and removes individual implementations from SharePoint, WebDav, Dropbox, Box, OneDrive, and GoogleDrive providers. The method is now declared in the base class ThirdPartyProviderDao and overridden where necessary. It also makes minor changes to the IsExistAsync method parameters in SharePointFileDao and SharePointFolderDao.