Implement URL shortening for confirmation links
The URL shortener has been applied to various confirmation links that are sent in notifications throughout the application. This was done to make the emails more user-friendly and to prevent long and complex URLs from being displayed in the emails. Instead of the standard URL, a shortened link is now provided.