Adjust file and folder copy/move conflict resolution
This commit improves the conflict resolution logic when moving or copying files and folders. Instead of checking for conflicts against the ProviderEntry field only, it now also checks the FileConflictResolveType. This change helps to better handle occurrences of duplicate files or folders and to skip certain items as required by the conflict resolution type. Additionally, an empty folders check is included in FolderDao to avoid unnecessary operations.