Improve user search case sensitivity
An update has been made to the user search functionality to ensure case insensitivity. The updated search now trims and converts the search text to lowercase and compares it with user details (first name, last name, title, location, email) that have been converted to lowercase as well. This adjustment allows for more accurate results regardless of the case used in the search term.